feat: add --duration flag and fix fps handling
- Add --duration flag to specify video length in seconds; resolves to
  nearest valid 8*k+1 frame count using ceil() to cover full duration
- Change output_fps from int to float throughout, defaulting to NATIVE_FPS
- Pass output_fps through to all pipeline configs instead of hardcoding
- Remove motion interpolation (minterpolate) from save functions; fps now
  directly controls generation rate, not post-processing
- Fix stale comment on default num_frames

+++ LTX_2_MLX/types.py
@@ -8,7 +8,7 @@ import mlx.core as mx
 
 # LTX-2 native frame rate. Matches frame_rate in the official Lightricks PipelineParams.
 # See: https://github.com/Lightricks/LTX-2/blob/main/packages/ltx-pipelines/src/ltx_pipelines/utils/constants.py
-NATIVE_FPS: int = 24
+NATIVE_FPS: float = 24.0
 
 class VideoPixelShape(NamedTuple):
     """
